requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.15.1,<0.16
Pillow>=9.1.0,<9.2
# uwsgi>=2.0.20,<2.1

# !!!
# In a Django project, adding packages to your requirements.txt file and installing
# them using pip is just the first step. This makes the packages available in your
# Python environment, but it doesn't mean that Django is aware of them or knows to use
# them in your application.

# When you add 'rest_framework' and 'drf_spectacular' to the INSTALLED_APPS
# list in your settings.py file, you're explicitly telling Django to include these
# applications in your project. This will enable Django to utilize their functionality,
# and allow it to find and use related modules, templates, static files, management commands, etc.